+++ DRN.Framework.EntityFramework/Context/DrnContextDevelopmentConnection.cs
@@ -13,8 +13,8 @@ public static class DrnContextDevelopmentConnection
             connectionString = devConnectionString;
         else
         {
-            var host = appSettings.Configuration.GetValue(DbContextConventions.DevHostKey, "postgresql");
-            var port = appSettings.Configuration.GetValue(DbContextConventions.DevPortKey, 5432);
+            var host = appSettings.Configuration.GetValue(DbContextConventions.DevHostKey, DbContextConventions.DefaultHost);
+            var port = appSettings.Configuration.GetValue(DbContextConventions.DevPortKey, DbContextConventions.DefaultPort);
             var username = appSettings.Configuration.GetValue<string>(DbContextConventions.DevUsernameKey, DbContextConventions.DefaultUsername);
             var database = appSettings.Configuration.GetValue<string>(DbContextConventions.DevDatabaseKey, DbContextConventions.DefaultDatabase);
             var password = appSettings.Configuration.GetValue<string>(DbContextConventions.DevPasswordKey);